Remove the openSUSE image
This was an attempt at running the tests on openSUSE, which we didn't
follow through with.

Since nothing use this, and none of us knows much about the openSUSE
release cycle (thus how to upgrade that image to newer openSUSE
releases), let's drop it for now.

Of course we'll be happy to bring it back in the future and run the
libcangjie/pycangjie/ibus-cangjie continuous integration on openSUSE.

See #1